Jparepository  id 제네릭 변경 user 도메인 생성 user repository생성

+++ src/main/java/com/gukbit/repository/BoardRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Long> {
+public interface BoardRepository extends JpaRepository<Board, Integer> {
 }
